requirements.txt
streamlit>=1.30.0
ultralytics>=8.0.150
torch>=2.0.0
opencv-python-headless>=4.8.0
Pillow
